examples/text_to_agent/web_search/gemini_background/pyproject.toml
[project]
name = "web-search-gemini-background"
version = "0.1.0"
description = "Web search agent using Gemini Background Search"
dependencies = [
    "cartesia-line",
    "google-genai",
    "pydantic",
    "loguru",
    "fastapi",
    "uvicorn",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
